refactor endpoints, UserController/Service and add ColorRoomSetController/Service, add GetMapping getUserById
