+++ backend/src/main/java/nilsot1/backend/controller/UserController.java
@@ -1,15 +1,11 @@
 package nilsot1.backend.controller;
 
 import lombok.RequiredArgsConstructor;
-import nilsot1.backend.exception.ColorRoomSetNotFoundException;
 import nilsot1.backend.exception.UserNotFoundException;
-import nilsot1.backend.model.ColorPalette;
-import nilsot1.backend.model.ColorRoomSet;
 import nilsot1.backend.model.User;
 import nilsot1.backend.service.UserService;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/user")
@@ -19,13 +15,8 @@ public class UserController {
     private final UserService service;
 
     @GetMapping("/{userId}")
-    public List<ColorRoomSet> getAllColorRoomSets(@PathVariable String userId) throws UserNotFoundException {
-        return service.getAllColorRoomSets(userId);
-    }
-
-    @GetMapping("/{userId}/colorRoomSets/{colorRoomSetId}")
-    public ColorRoomSet getColorRoomSetById(@PathVariable String userId, @PathVariable String colorRoomSetId) throws UserNotFoundException, ColorRoomSetNotFoundException {
-        return service.getColorRoomSetById(userId, colorRoomSetId);
+    public User getUserById(@PathVariable String userId) throws UserNotFoundException {
+        return service.getUserById(userId);
     }
 
     @PostMapping
@@ -33,14 +24,5 @@ public class UserController {
         return service.createNewUser(user);
     }
 
-    @PostMapping("/{userId}")
-    public User saveNewColorRoomSet(@PathVariable String userId, @RequestBody ColorRoomSet colorRoomSet) throws UserNotFoundException {
 
-        return service.saveNewColorRoomSet(colorRoomSet, userId);
-    }
-
-    @PutMapping("/{userId}/updateColorRoomSet/{colorRoomSetId}")
-    public User updateColorRoomSet(@PathVariable String userId, @PathVariable String colorRoomSetId, @RequestBody ColorPalette colorPalette) throws UserNotFoundException {
-        return service.updateColorPalette(userId, colorRoomSetId, colorPalette);
-    }
 }
